Create a basic Django application with TastyPie API
Summary: This patch adds a Django project and a core application. The data model is set up as Django models, and some models are exposed read-only with REST using TastyPie API. A rudimentary admin interface with basic model administration is also provided.

Test Plan: Read the code, run the Django application and check if it starts.

Reviewers: #modelstatus, havardf

Reviewed By: #modelstatus, havardf

Subscribers: havardf

Projects: #ecdiss

Differential Revision: https://phab.met.no/D1100