Move frontend web application code into Django views
Summary: This patch replaces the AngularJS web frontend with a simple non-dynamic page, and uses Django class views to generate the HTML. The responsibility to be a nice frontend is placed in the hands of Django admin.

Test Plan: Navigate to http://localhost:8000 and check that you get a data instance view.

Reviewers: #ecdiss, havardf

Reviewed By: #ecdiss, havardf

Subscribers: havardf

Projects: #ecdiss

Maniphest Tasks: T1942

Differential Revision: https://phab.met.no/D1281