Refactor message publishing logic
This commit splits the logic for model saves and sending messages on the
Kafka queue. The old functionality was to group the save transaction
with message sending, so that saving the object guarantees that a
message has been sent out. The database commit happens a little while
later, spawning problems with asynchronicity so that clients might
receive messages that were not to be found in the database.

In this commit, this problem is addressed by saving messages to a
dedicated message queue. However, this requires that a separate daemon
for sending Kafka messages is running.

The daemon polls the database for new messages, and can be started with:

    ./manage.py publisher